127 add comprehensive testing for the current bb parsing functionality (#135)
* Move file

* Update import path on SUT

* Import function that needs testing

* Add describe block for method

* Copy all existing tests from getTeams

* Uncomment and fix first test

* Fix Team construction warning

* Uncomment and update new test about parsing status

* Uncomment and rework test for eviction following understood BB25 model

* Add a test for whitespace and remove TODO comment

* Uncomment and rework test for expelled status

* Add test for runner-up status

* Add test w/ disclamer

* Uncomment and replace unnecessary test with stub for a new test for winner

* Uncomment and rework test to work and make appropriate assertion

* Uncomment and update test to make appropriate assertion

* Remove unnecessary test

* Fill out test about the winner

* Add a test to cover the runner up being ordered after the final evicted guest

* Add a test for the currentlyParticipating

* Downgrade these because they are a common occurrence

* Add a test to assert Blue's case

* Update test to make some stronger assertions
